+++ src/main/java/com/zeroToHero/accountingapp/dto/ReportDTO.java
@@ -11,8 +11,8 @@ import java.math.BigDecimal;
 @Data
 public class ReportDTO {
     private String name;
-    private BigDecimal purchasedQty;
-    private BigDecimal soldQty;
+    private Integer purchasedQty;
+    private Integer soldQty;
     private BigDecimal totalCost;
     private BigDecimal totalIncome;
