+++ src/main/java/com/zeroToHero/accountingapp/dto/StockDetailsDTO.java
@@ -1,10 +1,8 @@
 package com.zeroToHero.accountingapp.dto;
 
-import com.zeroToHero.accountingapp.entity.Product;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
@@ -17,8 +15,8 @@ public class StockDetailsDTO {
 
     private Long id;
     private LocalDateTime iDate;
-    private BigDecimal quantity;
+    private BigInteger quantity;
     private BigDecimal price;
-    private BigDecimal remainingQuantity;
-    private BigDecimal productId;
+    private BigInteger remainingQuantity;
+    private BigInteger productId;
 }
