+++ src/main/java/com/zeroToHero/accountingapp/entity/Product.java
@@ -1,6 +1,5 @@
 package com.zeroToHero.accountingapp.entity;
 
-
 import com.zeroToHero.accountingapp.enums.ProductStatus;
 import com.zeroToHero.accountingapp.enums.Unit;
 import lombok.Getter;
@@ -9,7 +8,6 @@ import lombok.Setter;
 import org.hibernate.annotations.Where;
 
 import javax.persistence.*;
-import java.math.BigDecimal;
 import java.math.BigInteger;
 
 @Entity
@@ -20,34 +18,30 @@ import java.math.BigInteger;
 public class Product extends BaseEntity {
 
     private String name;
-
     private String description;
-
 
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "category_id")
-   private Category category;
+    private Category category;
 
-
-    private BigDecimal qty;
+    private BigInteger qty;
 
     @Enumerated(EnumType.STRING)
     private Unit unit;
 
-    private BigDecimal lowLimitAlert;
-    private BigDecimal tax;
+    private BigInteger lowLimitAlert;
+    private BigInteger tax;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "company_id")
     private Company company;
-
 
     private Boolean enabled;
 
     @Enumerated(EnumType.STRING)
     private ProductStatus productStatus;
 
-    private BigDecimal newColumn;
+    private BigInteger newColumn;
 
 }
